Hi Sneer
I am the author of a project called Sneer. :)

Please take a look: http://sneer.me/about.html

It would mean a lot for us if we could use the 'sneer' username here on github. It seems to be quite inactive.

Could you donate it to us?

We are developing a github-based code sharing tool and it would mean the world to us to be able to base our own code on the sneer user.

Thank you very much!

Klaus

PS: A pull-request was the only way I found to get in touch :)